Move structure models to dedicated folder
