+++ LoxNet.Client/Models/KeyInfo.cs
@@ -1,0 +1,3 @@
+namespace LoxNet;
+
+public record KeyInfo(string Key, string Salt, string HashAlg);
