+++ LoxNet.Client/Models/LoxoneCategory.cs
@@ -1,0 +1,3 @@
+namespace LoxNet;
+
+public record LoxoneCategory(string Id, string Name, string? Type, string? Color);
