+++ LoxNet.Client/Models/LoxoneMessage.cs
@@ -1,0 +1,5 @@
+namespace LoxNet;
+
+using System.Text.Json;
+
+public record LoxoneMessage(int Code, JsonElement? Value, string? Message);
