+++ LoxNet.Client/Models/LoxoneRoom.cs
@@ -1,0 +1,3 @@
+namespace LoxNet;
+
+public record LoxoneRoom(string Id, string Name, string? Image, int? DefaultRating);
