+++ LoxNet.Client/Models/TokenInfo.cs
@@ -1,0 +1,3 @@
+namespace LoxNet;
+
+public record TokenInfo(string Token, long ValidUntil, int TokenRights, bool UnsecurePass, string Key);
